add the package dist files to git ignore and rename to sensorAtlas
