+++ .gitignore
@@ -1,2 +1,4 @@
 venv
 *.idea
+dist
+sensorAtlas.egg-info
