--- pymatau.egg-info/SOURCES.txt
@@ -1,60 +1,0 @@
-LICENSE
-MANIFEST.in
-README.md
-setup.py
-pymatau/__init__.py
-pymatau/apps.py
-pymatau/errors.py
-pymatau/models.py
-pymatau/query_options.py
-pymatau/representation.py
-pymatau/serializer.py
-pymatau/urls.py
-pymatau/views.py
-pymatau/viewsets.py
-pymatau.egg-info/PKG-INFO
-pymatau.egg-info/SOURCES.txt
-pymatau.egg-info/dependency_links.txt
-pymatau.egg-info/requires.txt
-pymatau.egg-info/top_level.txt
-pymatau/routers/__init__.py
-pymatau/routers/datastream.py
-pymatau/routers/featureofinterest.py
-pymatau/routers/historicallocation.py
-pymatau/routers/location.py
-pymatau/routers/observation.py
-pymatau/routers/observedproperty.py
-pymatau/routers/router.py
-pymatau/routers/sensor.py
-pymatau/routers/thing.py
-pymatau/routers/__pycache__/__init__.cpython-37.pyc
-pymatau/routers/__pycache__/datastream.cpython-37.pyc
-pymatau/routers/__pycache__/featureofinterest.cpython-37.pyc
-pymatau/routers/__pycache__/historicallocation.cpython-37.pyc
-pymatau/routers/__pycache__/location.cpython-37.pyc
-pymatau/routers/__pycache__/observation.cpython-37.pyc
-pymatau/routers/__pycache__/observedproperty.cpython-37.pyc
-pymatau/routers/__pycache__/router.cpython-37.pyc
-pymatau/routers/__pycache__/sensor.cpython-37.pyc
-pymatau/routers/__pycache__/thing.cpython-37.pyc
-tests/.DS_Store
-tests/__init__.py
-tests/models.py
-tests/test_complianceA1.py
-tests/test_complianceA2.py
-tests/test_complianceA3.py
-tests/test_complianceA4.py
-tests/test_complianceA6.py
-tests/test_settings.py
-tests/urls.py
-tests/wsgi.py
-tests/__pycache__/__init__.cpython-37.pyc
-tests/__pycache__/models.cpython-37.pyc
-tests/__pycache__/test_complianceA1.cpython-37.pyc
-tests/__pycache__/test_complianceA2.cpython-37.pyc
-tests/__pycache__/test_complianceA3.cpython-37.pyc
-tests/__pycache__/test_complianceA4.cpython-37.pyc
-tests/__pycache__/test_complianceA6.cpython-37.pyc
-tests/__pycache__/test_settings.cpython-37.pyc
-tests/__pycache__/test_unit.cpython-37.pyc
-tests/__pycache__/urls.cpython-37.pyc
