--- pymatau/__init__.py
@@ -1,2 +1,0 @@
-# app config
-default_app_config = 'pymatau.apps.SensorAppConfig'
