+++ sensorAtlas/__init__.py
@@ -1,0 +1,2 @@
+# app config
+default_app_config = 'sensorAtlas.apps.SensorAppConfig'
