+++ sensorAtlas/apps.py
@@ -3,5 +3,5 @@ from django.utils.translation import ugettext_lazy as _
 
 
 class SensorAppConfig(AppConfig):
-    name = 'pymatau'
+    name = 'sensorAtlas'
     verbose_name = _('Sensors')
