+++ sensorAtlas/errors.py
