+++ sensorAtlas/views.py
@@ -1,6 +1,6 @@
 from .models import Location, Thing, DataStream, Sensor, \
     ObservedProperty, Observation, FeatureOfInterest, HistoricalLocation
-import pymatau.serializer as serializers
+import sensorAtlas.serializer as serializers
 from .query_options import Filter, Orderby
 from .viewsets import ViewSet, PropertyPath
 from rest_framework import generics
