+++ setup.py
@@ -8,9 +8,9 @@ with open(os.path.join(os.path.dirname(__file__), 'README.md')) as readme:
 os.chdir(os.path.normpath(os.path.join(os.path.abspath(__file__), os.pardir)))
 
 setup(
-    name='pymatau',
+    name='sensorAtlas',
     version='0.1',
-    packages=['pymatau'],
+    packages=['sensorAtlas'],
     include_package_data=True,
     license='MIT License',
     description='An OGC SensorThings API implementation in Python.',
